@@ -5,7 +5,7 @@
   <modelVersion>4.0.0</modelVersion>
 
   <groupId>com.dmst.president</groupId>
-  <artifactId>social-network</artifactId>
+  <artifactId>UniPost</artifactId>
   <version>1.0-SNAPSHOT</version>
   <packaging>jar</packaging>
 
@@ -15,16 +15,16 @@
 
   <properties>
     <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-    <maven.compiler.source>1.7</maven.compiler.source>
-    <maven.compiler.target>1.7</maven.compiler.target>
+    <maven.compiler.source>1.8</maven.compiler.source>
+    <maven.compiler.target>1.8</maven.compiler.target>
   </properties>
 
   <dependencies>
-    <dependency> <!--JUnit-->
-      <groupId>junit</groupId>
-      <artifactId>junit</artifactId>
-      <version>4.11</version>
-      <scope>test</scope>
+    <dependency> <!--JUnit5 jupiter-->
+      <groupId>org.junit.jupiter</groupId>
+        <artifactId>junit-jupiter-engine</artifactId>
+        <version>5.4.0</version>
+        <scope>test</scope>
     </dependency>
     <dependency> <!--Checkstyle-->
       <groupId>com.puppycrawl.tools</groupId>
@@ -32,7 +32,7 @@
       <version>10.4</version>
     </dependency>
     
-    <dependency>
+    <dependency> <!--Spotbugs-->
       <groupId>com.github.spotbugs</groupId>
       <artifactId>spotbugs-maven-plugin</artifactId>
       <version>4.7.3.0</version>
@@ -57,7 +57,7 @@
           </reportSets>
         </plugin>
 
-        <plugin>
+        <plugin> <!--Spotbugs-->
                 <groupId>com.github.spotbugs</groupId>
                 <artifactId>spotbugs-maven-plugin</artifactId>
                 <version>3.1.8</version>
@@ -85,13 +85,37 @@
           <artifactId>maven-compiler-plugin</artifactId>
           <version>3.8.0</version>
         </plugin>
+        <!-- Unit tests using Junit-->
         <plugin>
           <artifactId>maven-surefire-plugin</artifactId>
           <version>2.22.1</version>
         </plugin>
         <plugin>
+        <groupId>org.apache.maven.plugins</groupId>
+        <!-- Integration tests using Junit. Might be unnecessary for our project-->
+        <artifactId>maven-failsafe-plugin</artifactId>
+        <version>3.0.0-M7</version>
+        <executions>
+          <execution>
+            <goals>
+              <goal>integration-test</goal>
+              <goal>verify</goal>
+            </goals>
+          </execution>
+        </executions>
+      </plugin>
+        <plugin>
           <artifactId>maven-jar-plugin</artifactId>
           <version>3.0.2</version>
+          <!-- Configure jar plugin to state main classpath-->
+          <configuration>
+            <archive>
+              <manifest>
+                <addClasspath>true</addClasspath>
+                <mainClass>com.dmst.president.UniPost</mainClass>
+             </manifest>
+           </archive>
+         </configuration>
         </plugin>
         <plugin>
           <artifactId>maven-install-plugin</artifactId>
@@ -134,7 +158,7 @@
             </executions>
           </plugin>
 
-          <plugin>
+          <plugin> <!--Spotbugs-->
             <groupId>com.github.spotbugs</groupId>
             <artifactId>spotbugs-maven-plugin</artifactId>
             <version>4.7.2.1</version>
